fix: address CodeRabbit review on PR #1 (safe fixes + tick-size bug + flatten gate)
Resolves the actionable CodeRabbit findings that are safe/correct now, plus the
HIGH tick-size quantization live-blocker and a sign-off'd account-wide-flatten
gate. The two big CRITICALs (get_position restart reconciliation, flatten() 207
multi-status) stay deferred in the CLAUDE.md backlog by design.

Engine / pure core
- Quantize Setup stop & target to tick_size in _setup_from, AWAY from entry
  (LONG stop floors / target ceils; SHORT inverted), before rr is recomputed so
  validate_setup's RR floor stays consistent. Uses ROUND_FLOOR/ROUND_CEILING so
  the direction is correct regardless of sign. Fixes Alpaca sub-penny TP/SL
  bracket rejection when the ATR buffer dominates. (engine._quantize_price)
- Add Engine.current_atr() -> Decimal | None (None before seeded, no raise).

Execution / orchestrator
- Account-wide flatten safety gate: new Broker.list_position_symbols(); in LIVE
  mode _preflight aborts (CRITICAL log + RuntimeError) if the account holds any
  non-run.symbol position, so the account-wide EOD flatten can't liquidate an
  unrelated holding.
- _revalidate_setup uses engine.current_atr() (drops fragile getattr ctx sniff).
- whole_share_qty rejects non-int/bool (TypeError) before the <1 check.
- trade-updates Queue bounded at maxsize=1024 (block-on-full back-pressure).
- ApprovalRequest.risk_dollars float -> Decimal (money is never float).

Config / indicators / logging
- find_impulse takes body_ratio/location params; detect_displacement passes
  cfg.strong_close_body_ratio/strong_close_location (defaults preserve behavior).
- RunConfig.symbol normalized (strip().upper()) so order submission and the
  flatten gate agree regardless of how it's typed; reject empty.
- Validate tick_size > 0 (fail fast vs DivisionByZero in quantization).
- logconf TimeStamper utc=True (Z-suffixed, unambiguous timestamps).
- Reconcile CLAUDE.md Python version to 3.11+ (matches pyproject/README).

Tests
- Remove the tests/.context shim entirely (delete tests/context.py; direct
  editable-package imports per project convention); drop a stale type: ignore;
  tighten the purity regex to catch dt.now(.
- New: test_engine_quantization.py; find_impulse threshold, whole_share_qty
  type, list_position_symbols, flatten-gate, and config-validator tests.

Gate green: ruff + mypy + lint-imports clean; 415 passed, 7 skipped, 0 errors.
Adversarially verified via a 4-agent review workflow before commit.

+++ orb_bot/models.py
@@ -120,7 +120,7 @@ class ApprovalRequest:
     setup: Setup
     symbol: str
     qty: int
-    risk_dollars: float
+    risk_dollars: Decimal
     mode: str                    # 'PAPER' | 'LIVE'
     feed: str
     or_high: Decimal
